orm mapping missing SonataMediaBundle
I had a lot of trouble with this: for the doctrine.orm.entity_managers.default.mappings section it is said to add ApplicationSonataMediaBundle, but then, when you update the database schema, you get tables with only an 'id' field. So, after much googling I found out that you should also add SonataMediaBundle, and also your own bundles, because somehow the auto_mapping does not work anymore.