Fix failing regexp tests on output of CleanMediaCommand
The order of output of CleanMediaCommand is non-deterministic,
which makes the regexp tests fail sometimes.

The regexp used to match the output was expected to contains:
  - first,  information about qwertz.ext
  - second, information about thumb_1_bar.ext

However, the order of the output is not important to us.
The real semantic of the test is to check that both
informations are present, order is irrelevant.

The new assertion checks for both informations regardless of the order.