Ensure that repository calculations work with UUIDs
Here media::id is used for some calculations to define a repository foldername. If media::id is UUID (string) instead of integer, this calculation will throw an exception in PHP 7.1 becauce math operations on strings are not allowed.

This commit uses the first two segments of UUID and converts their hex value to decimal to use as base of calculation.