Fixed deletion of images with id greater than 1000
Inside the $this->getReferenceImage() method, the id of the $media object is used, but at the time of the postRemove event, the $media object is detached from an EntityManager and its id is null. During the postRemove event, the $media->getId() method returns null, and the $this->getReferenceImage() method returns the wrong path and the file remains in the file system as garbage. My changes fix this.