Fixed deletion of images with id greater than 1000 (#1592)
* Fixed deletion of images with id greater than 1000

Inside the $this->getReferenceImage() method, the id of the $media object is used, but at the time of the postRemove event, the $media object is detached from an EntityManager and its id is null. During the postRemove event, the $media->getId() method returns null, and the $this->getReferenceImage() method returns the wrong path and the file remains in the file system as garbage. My changes fix this.

* Add testPostRemove().

Test for "Fix deletion of images with id greater than 1000".

* Update testPostRemove().

* The visibility of the property "clones" is changed to private.

* Adapted test "testPostRemove" to check the data in the private property "clones" instead of the protected.

* Update src/Provider/BaseProvider.php

Co-Authored-By: Javier Spagnoletti <phansys@gmail.com>

* Update src/Provider/BaseProvider.php

Co-Authored-By: Javier Spagnoletti <phansys@gmail.com>

* Update tests/Provider/BaseProviderTest.php

Co-Authored-By: Javier Spagnoletti <phansys@gmail.com>

* Adding a separate version of the testPostRemove(), which can be run in the current branch 3, and in the proposed me.

* Revert "Adding a separate version of the testPostRemove(), which can be run in the current branch 3, and in the proposed me."

This reverts commit bab31c8e3b3d662651634a4d90ccdd456adddde1.

* Adding a separate version of the testPostRemove(), which can be run in the current branch 3, and in the proposed me.

* Remove unnecessary comment.

* Removed unnecessary comment

* Reorganization of tests for compliance with the requirements.

* Update tests/Provider/BaseProviderTest.php

Co-Authored-By: Javier Spagnoletti <phansys@gmail.com>

* Fixes for FlintCI.

* Fixed deletion of images with id greater than 1000

Inside the $this->getReferenceImage() method, the id of the $media object is used, but at the time of the postRemove event, the $media object is detached from an EntityManager and its id is null. During the postRemove event, the $media->getId() method returns null, and the $this->getReferenceImage() method returns the wrong path and the file remains in the file system as garbage. My changes fix this.

* Revert "Fixed deletion of images with id greater than 1000"

This reverts commit a74315c5ca2b4a8b614b35c568d7c65efed28667.