fix(thumbnail): add 'addResizer' method call only if the method exists
The ThumbnailCompilerPass add the "addResizer" method call on the thumbnail service
but this method is not on the ThumbnailInterface, so there is no guarantee it actually exists.
I do not take into account the fact that the service could be a child definition, as this is not the
case in "our" implementation. If someone alters it, it is his/her responsability.