Fix CDN flush behavior
The CDN flush behavior is an asynchronous operation. It consists in telling the CDN a path must be invalidated,
in order to pick the path content again from its origin. This operation can take some time, so the invalidation
result must be obtained using long polling, a scheduled task (like `UpdateCdnStatusCommand`) or something similar,
in order to know if it has finished successfully or not. When this operation is done, the result is passed to `MediaInterface::setCdnStatus()`.
The users are able to mark a medium as flushable (indicating it must be track any update and kept fresh in the CDN)
by calling MediaInterface::setCdnIsFlushable().

Previous to this change, the behavior at `BaseProvider::flushCdn()` and UpdateCdnStatusCommand were making wrong assumptions
about of the purpose of `MediaInterface::getCdnIsFlushable()`. They were calling `MediaInterface::setCdnIsFlushable()` arbitrarily,
but they should not do that since it was overriding what the users have chosen about the CDN content for each medium.

Additionally to this fix, these changes also allow `BaseProvider::flushCdn()` to update the flush status of a medium
with a pending flush, if its result is successful. It also prevents to attempt a new flush request if a previous one
is in a pending (different than `CDNInterface::STATUS_OK` or `CDNInterface::STATUS_ERROR`) status.